Main group 33 css

+++ src/components/ProductCard.jsx
@@ -11,19 +11,33 @@ class ProductCard extends React.Component {
   addCartItem({ id, title, price, inStock }) {
     const evaluation = [];
     const cartItemProperties = { id, title, price, inStock, evaluation };
-    cartItemProperties.quantity = 1;
+    const { onAddCard } = this.props;
     if (!localStorage.cartItems) {
+      cartItemProperties.quantity = 1;
       localStorage.setItem('cartItems', JSON.stringify([cartItemProperties]));
+      onAddCard();
     } else {
       const itemsInStorage = localStorage.getItem('cartItems');
       const parsedItems = JSON.parse(itemsInStorage);
-      localStorage.setItem(
-        'cartItems',
-        JSON.stringify(parsedItems.concat(cartItemProperties)),
-      );
+      const productIsAlreadyInCart = parsedItems
+        .some((item) => (item.id === cartItemProperties.id));
+      if (productIsAlreadyInCart) {
+        const index = parsedItems
+          .findIndex((item) => item.id === cartItemProperties.id);
+        parsedItems[index].quantity += 1;
+        localStorage.setItem(
+          'cartItems',
+          JSON.stringify(parsedItems),
+        );
+      } else {
+        cartItemProperties.quantity = 1;
+        localStorage.setItem(
+          'cartItems',
+          JSON.stringify(parsedItems.concat(cartItemProperties)),
+        );
+        onAddCard();
+      }
     }
-    const { onAddCard } = this.props;
-    onAddCard();
   }
 
   render() {
@@ -32,32 +46,36 @@ class ProductCard extends React.Component {
     const { title, thumbnail, price, id, shipping } = product;
     const p = <p data-testid="free-shipping">Frete grátis!</p>;
     return (
-      <div data-testid="product">
-        <Link
-          to={ {
-            pathname: `/details/${id}/category/${product.category_id}`,
-          } }
-          data-testid="product-detail-link"
-        >
-          <h5>{ title }</h5>
-        </Link>
+      <div data-testid="product" className="cardProduct">
         <img src={ thumbnail } alt="product item" />
-        <div>
-          <h6>
-            R$
-            {price}
-          </h6>
-        </div>
-        <div>
-          {shipping.free_shipping ? p : ''}
+        <div className="product-container">
+          <Link
+            className="product-detail-link"
+            to={ {
+              pathname: `/details/${id}/category/${product.category_id}`,
+            } }
+            data-testid="product-detail-link"
+          >
+            <h5>{title}</h5>
+            <div>
+              <h6>
+                R$
+                { price }
+              </h6>
+            </div>
+            <div>
+              {shipping.free_shipping ? p : ''}
+            </div>
+          </Link>
         </div>
         <button
           type="button"
           name="productId"
           data-testid="product-add-to-cart"
+          className="product-add-to-cart"
           onClick={ () => this.addCartItem({ id, title, price, inStock }) }
         >
-          Adicionar ao Carinho
+          Adicionar
         </button>
       </div>
     );
